tl: use google cloud API
The reverse-engineered "API" used by the web frontend was completely
broken so this plugin was useless until now. Switch to the official API
which has a free tier with a very reasonable quota.

The plugin now has an allow-list so that it will only be enabled in
blessed groups when running as a bot.

You need to set up a google cloud account, enable translate and TTS APIs
and create a service account key for this to work. Save the key as
google_cloud_key.json in the root of your borg tree.